+++ angular/lib/src/source_gen/template_compiler/find_components.dart
@@ -438,8 +438,6 @@ class ComponentVisitor
               propertyType is ParameterizedType &&
               _htmlElement
                   .isAssignableFromType(propertyType.typeArguments.first),
-      isQueryListType: propertyType?.element != null &&
-          $QueryList.isExactlyType(propertyType),
       read: readType != null
           ? new CompileTokenMetadata(
               identifier: new CompileIdentifierMetadata(